Create interface+abstract class for scheduler builders.
